fix(toolbox): authenticate the registry's fetch too — the one the e2e actually uses
The kind e2e drives the registry's install capabilities, wired to the
plain http_get in crates/registry — not the desktop's progress fetch
the first fix patched, so the 403 flake would have persisted. The
predicate and ambient-token read move into kube's toolbox_install
(shared by both crates), and BOTH blocking fetch paths consult that one
predicate so they can't drift on which hosts get the token.
